requirements.txt
# Core dependencies
streamlit==1.29.0
openai==1.3.7
python-dotenv==1.0.0

# PDF generation (reportlab works on all platforms)
reportlab==4.0.7

# PDF reading - using pypdf instead of PyMuPDF for Windows compatibility
pypdf==3.17.4
pdf2image==1.16.3

# OCR support (optional - will work without it)
pytesseract==0.3.10
Pillow==10.1.0

# Data handling
pandas==2.1.4
numpy==1.26.2

# Audio processing for voice input
SpeechRecognition==3.10.1
pydub==0.25.1

# WhatsApp integration
twilio==8.10.3

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pydantic==2.5.2

# Optional but recommended
streamlit-extras==0.3.6
watchdog==3.0.0

# For medical validation
difflib2==0.1.0